requirements.txt
fastapi~=0.109.2
starlette~=0.36.3
pydantic~=2.6.1
redis~=5.0.1
aiocache~=0.12.2
python-dotenv~=1.0.1
pydantic-settings~=2.1.0
snowflake-sqlalchemy~=1.5.1
hypercorn[standard]~=0.16.0
fastapi-azure-auth~=4.3.0
openai==1.13.3
pandas==2.2.1
numpy~=1.26.4
pytz~=2024.1
python-dateutil~=2.8.2
vertexai~=0.0.1
matplotlib~=3.8.2
seaborn~=0.13.2
fastembed~=0.2.1
msal~=1.26.0
tiktoken~=0.6.0
sentence_transformers~=2.3.1
torch~=2.2.0
tabulate~=0.9.0
evaluate~=0.4.1
rouge-score~=0.1.2
bert_score~=0.3.13
openpyxl~=3.1.2
tqdm~=4.66.2
google-auth~=2.27.0
google-cloud-aiplatform~=1.41.0
boto3~=1.34.41
python-jose~=3.3.0
botocore~=1.34.41
scikit-learn~=1.4.0
requests~=2.31.0
asyncpg~=0.29.0
SQLAlchemy~=1.4.51
toml~=0.10.2
uv~=0.1.7
aiohttp~=3.9.3
PyJWT~=2.8.0
pytest~=8.1.1
pytest-dotenv~=0.5.2
pytest-asyncio~=0.23.6
opentelemetry-distro==0.44b0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-logging==0.44b0
opentelemetry-instrumentation-sqlalchemy==0.44b0
opentelemetry-instrumentation-openai==0.13.2
opentelemetry-sdk-extension-aws==2.0.1
opentelemetry-propagator-aws-xray==1.0.1
